apps/agent/requirements.txt
aiofiles==23.2.1 ; python_version >= "3.11" and python_version < "4.0"
annotated-types==0.7.0 ; python_version >= "3.11" and python_version < "4.0"
anyio==4.7.0 ; python_version >= "3.11" and python_version < "4.0"
async-timeout==5.0.1 ; python_version >= "3.11" and python_full_version < "3.11.3"
audioop-lts==0.2.1 ; python_version >= "3.13" and python_version < "4.0"
beautifulsoup4==4.12.3 ; python_version >= "3.11" and python_version < "4.0"
bs4==0.0.2 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.0 ; python_version >= "3.11" and python_version < "4.0"
certifi==2024.12.14 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.0 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
distro==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
dnspython==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
email-validator==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
fastapi-cli[standard]==0.0.7 ; python_version >= "3.11" and python_version < "4.0"
fastapi==0.115.6 ; python_version >= "3.11" and python_version < "4.0"
fastapi[all]==0.115.6 ; python_version >= "3.11" and python_version < "4.0"
ffmpy==0.5.0 ; python_version >= "3.11" and python_version < "4.0"
filelock==3.16.1 ; python_version >= "3.11" and python_version < "4.0"
fsspec==2024.12.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.24.0 ; python_version >= "3.11" and python_version < "4.0"
google-api-python-client==2.156.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.37.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.66.0 ; python_version >= "3.11" and python_version < "4.0"
gradio-client==1.5.2 ; python_version >= "3.11" and python_version < "4.0"
gradio==5.9.1 ; python_version >= "3.11" and python_version < "4.0"
greenlet==3.1.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio-tools==1.68.1 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.68.1 ; python_version >= "3.11" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.11" and python_version < "4.0"
h2==4.1.0 ; python_version >= "3.11" and python_version < "4.0"
hpack==4.0.0 ; python_version >= "3.11" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.11" and python_version < "4.0"
httplib2==0.22.0 ; python_version >= "3.11" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.11" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
httpx[http2]==0.28.1 ; python_version >= "3.11" and python_version < "4.0"
huggingface-hub==0.27.0 ; python_version >= "3.11" and python_version < "4.0"
hyperframe==6.0.1 ; python_version >= "3.11" and python_version < "4.0"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.5 ; python_version >= "3.11" and python_version < "4.0"
jiter==0.8.2 ; python_version >= "3.11" and python_version < "4.0"
markdown-it-py==3.0.0 ; python_version >= "3.11" and python_version < "4.0"
markdownify==0.14.1 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==2.1.5 ; python_version >= "3.11" and python_version < "4.0"
mdurl==0.1.2 ; python_version >= "3.11" and python_version < "4.0"
numpy==2.2.1 ; python_version >= "3.11" and python_version < "4.0"
openai==1.58.1 ; python_version >= "3.11" and python_version < "4.0"
orjson==3.10.12 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
pandas==2.2.3 ; python_version >= "3.11" and python_version < "4.0"
pillow==11.0.0 ; python_version >= "3.11" and python_version < "4.0"
playwright==1.49.1 ; python_version >= "3.11" and python_version < "4.0"
portalocker==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.25.0 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.2 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "4.0"
pydantic-extra-types==2.10.1 ; python_version >= "3.11" and python_version < "4.0"
pydantic-settings==2.7.0 ; python_version >= "3.11" and python_version < "4.0"
pydantic==2.10.4 ; python_version >= "3.11" and python_version < "4.0"
pydub==0.25.1 ; python_version >= "3.11" and python_version < "4.0"
pyee==12.0.0 ; python_version >= "3.11" and python_version < "4.0"
pygments==2.18.0 ; python_version >= "3.11" and python_version < "4.0"
pyparsing==3.2.0 ; python_version >= "3.11" and python_version < "4.0"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "4.0"
python-multipart==0.0.20 ; python_version >= "3.11" and python_version < "4.0"
pytz==2024.2 ; python_version >= "3.11" and python_version < "4.0"
pywin32==308 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
pyyaml==6.0.2 ; python_version >= "3.11" and python_version < "4.0"
qdrant-client==1.12.1 ; python_version >= "3.11" and python_version < "4.0"
redis==5.2.1 ; python_version >= "3.11" and python_version < "4.0"
regex==2024.11.6 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rich-toolkit==0.12.0 ; python_version >= "3.11" and python_version < "4.0"
rich==13.9.4 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4"
ruff==0.8.4 ; python_version >= "3.11" and python_version < "4.0" and sys_platform != "emscripten"
safehttpx==0.1.6 ; python_version >= "3.11" and python_version < "4.0"
semantic-version==2.10.0 ; python_version >= "3.11" and python_version < "4.0"
setuptools==75.6.0 ; python_version >= "3.11" and python_version < "4.0"
shellingham==1.5.4 ; python_version >= "3.11" and python_version < "4.0"
six==1.17.0 ; python_version >= "3.11" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.11" and python_version < "4.0"
soupsieve==2.6 ; python_version >= "3.11" and python_version < "4.0"
starlette==0.41.3 ; python_version >= "3.11" and python_version < "4.0"
tiktoken==0.8.0 ; python_version >= "3.11" and python_version < "4.0"
tomlkit==0.13.2 ; python_version >= "3.11" and python_version < "4.0"
tqdm==4.67.1 ; python_version >= "3.11" and python_version < "4.0"
typer==0.15.1 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
tzdata==2024.2 ; python_version >= "3.11" and python_version < "4.0"
ujson==5.10.0 ; python_version >= "3.11" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
uvicorn==0.32.1 ; python_version >= "3.11" and python_version < "4.0"
uvicorn[standard]==0.32.1 ; python_version >= "3.11" and python_version < "4.0"
uvloop==0.21.0 ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.11" and python_version < "4.0"
watchfiles==1.0.3 ; python_version >= "3.11" and python_version < "4.0"
websockets==14.1 ; python_version >= "3.11" and python_version < "4.0"
